+++ src/components/NcRichContenteditable/NcRichContenteditable.vue
@@ -861,6 +861,7 @@ export default {
 	&__input:focus + &__label,
 	&__input:not(&__input--empty) + &__label {
 		inset-block-start: -10px;
+		line-height: 1.5; // minimum allowed line height for accessibility
 		font-size: 13px; // minimum allowed font size for accessibility
 		font-weight: 500;
 		border-radius: var(--default-grid-baseline) var(--default-grid-baseline) 0 0;